added the auth to edit and delete in reviews controllers

+++ controllers/reviews.js
@@ -1,6 +1,6 @@
 const Review = require("../models/review.js");
 const Service = require("../models/service.js");
-const verifyToken = require("../middleware/verify-token.js");
+const verifyToken = require("../middleware/verify-token.js"); 
 const express = require("express");
 const router = express.Router();
 
@@ -56,23 +56,29 @@ router.get("/service/:serviceId", async (req, res) => {
 });
 
 // remove a review
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", verifyToken, async (req, res) => { 
   try {
     const reviewId = req.params.id;
-    const deleteReview = await Review.findByIdAndDelete(reviewId);
+    const review = await Review.findById(reviewId);
 
-    if (!deleteReview) {
+    if (!review) {
       res.status(404);
       throw new Error("Review not found.");
     }
 
-    const service = await Service.findById(deleteReview.service);
+    if (review.user.toString() !== req.user.id) { 
+      return res.status(403).json({ error: "Unauthorized: You are not the owner of this review." });
+    }
+
+    await Review.findByIdAndDelete(reviewId);
+
+    const service = await Service.findById(review.service);
     service.reviews = service.reviews.filter(
       (id) => id.toString() !== reviewId.toString()
     );
     await service.save();
 
-    res.status(200).json(deleteReview);
+    res.status(200).json({ message: "Review deleted successfully." });
   } catch (err) {
     if (res.statusCode === 404) {
       res.json({ error: err.message });
@@ -83,19 +89,25 @@ router.delete("/:id", async (req, res) => {
 });
 
 // edit the review
-router.put("/:id", async (req, res) => {
+router.put("/:id", verifyToken, async (req, res) => { 
   try {
     const reviewId = req.params.id;
+    const review = await Review.findById(reviewId);
+
+    if (!review) {
+      res.status(404);
+      throw new Error("Review not found.");
+    }
+
+    if (review.user.toString() !== req.user.id) { 
+      return res.status(403).json({ error: "Unauthorized: You are not the owner of this review." });
+    }
+
     const updatedReview = await Review.findByIdAndUpdate(reviewId, req.body, {
       new: true,
     })
       .populate("user")
       .populate("service");
-
-    if (!updatedReview) {
-      res.status(404);
-      throw new Error("Review not found.");
-    }
 
     res.status(200).json(updatedReview);
   } catch (err) {
